Fully separate File and Folder Controllers from DB calls, to make DB substitution easier
